feat: add floor
A floor component, built from an array of 2D points.
When created in a scene, the player will be forced to interpolate
between two floor points, aligned with the players feet.

The TMX scene loader was modified to load a new object of type 'floor';
a polygon shape that defines a floor component.